GSYE-426-state: Move state.py into package in strategy
