+++ src/gsy_e/models/strategy/energy_parameters/energy_params_eb.py
@@ -32,10 +32,10 @@ from gsy_framework.forward_markets.forward_profile import ForwardTradeProfileGen
 from gsy_framework.utils import convert_kW_to_kWh
 
 from gsy_e.constants import FLOATING_POINT_TOLERANCE, FORWARD_MARKET_MAX_DURATION_YEARS
-from gsy_e.models.state import LoadState, PVState
+from gsy_e.models.strategy.state import LoadState, PVState
 
 if TYPE_CHECKING:
-    from gsy_e.models.state import StateInterface
+    from gsy_e.models.strategy.state import StateInterface
     from gsy_e.models.area import AreaBase
