+++ src/gsy_e/models/strategy/load_hours.py
@@ -37,7 +37,7 @@ from gsy_e.gsy_e_core.exceptions import MarketException
 from gsy_e.gsy_e_core.util import get_market_maker_rate_from_config
 from gsy_e.models.base import AssetType
 from gsy_e.models.market import MarketBase
-from gsy_e.models.state import LoadState
+from gsy_e.models.strategy.state import LoadState
 from gsy_e.models.strategy import BidEnabledStrategy
 from gsy_e.models.strategy.energy_parameters.load import LoadHoursPerDayEnergyParameters
 from gsy_e.models.strategy.future.strategy import future_market_strategy_factory
